requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.3
Deprecated==1.2.13
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
iniconfig==1.1.1
kombu==5.2.4
linecache2==1.0.0
Markdown==3.3.7
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
redis==4.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
traceback2==1.4.0
unittest2==1.1.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
